+++ packages/ui/src/listen/minimalism/index.ts
@@ -1,9 +1,9 @@
 import MinimalismThemeProvider from './ThemeProvider';
 import MusicWidget from './music-widget';
 import { Header } from './header/header';
-import { FeelingList } from './homeContainer/feeling-list';
-import { AudioList } from './homeContainer/audio-list';
-import { MainContainer } from './homeContainer/main-container';
+import { FeelingList } from './home/feeling-list';
+import { AudioList } from './home/audio-list';
+import { MainContainer } from './home/main-container';
 
 export {
   MinimalismThemeProvider,
